+++ src/postulo/core/server_urls.py
@@ -11,6 +11,7 @@ urlpatterns = [
     path("people/<int:pk>/administrator/", views.PersonAdminView.as_view(), name="person_admin"),
     path("people/<int:pk>/active/", views.PersonActiveView.as_view(), name="person_active"),
     path("people/<int:pk>/username/", views.PersonUsernameView.as_view(), name="person_username"),
+    path("people/<int:pk>/delete/", views.PersonDeleteView.as_view(), name="person_delete"),
     path("sign-in/", views.SignInView.as_view(), name="signin"),
     path("email/", views.EmailView.as_view(), name="email"),
     path("email/test/", views.EmailTestView.as_view(), name="email_test"),